convert TouchDetection to class component Mixin
